feat(config): Add new package config
- Refactoring ServerSettingsProvider onto new package
- Add decorators @Value and @Constant that can be used on class property.
Theses decorators will take the value from globalServerSettings.
- Add globalServerSettings symbol.
