+++ src/config/services/ServerSettingsService.ts
@@ -1,0 +1,11 @@
+/**
+ * @module common/server
+ */
+/** */
+import {ServerSettingsProvider} from "../class/ServerSettingsProvider";
+
+/**
+ * `ServerSettingsService` contains all information about [ServerLoader](api/common/server/serverloader.md) configuration.
+ */
+export class ServerSettingsService extends ServerSettingsProvider {
+}
