+++ src/core/interfaces/Env.ts
@@ -1,8 +1,14 @@
 /**
- * @module common/core
- */ /** */
-
-/**
  *
  */
-export type Env = "production" | "development" | "test";
+export enum Env {
+    PROD = "production",
+    DEV = "development",
+    TEST = "test"
+}
+
+/**
+ * Fallback
+ * @type {Env}
+ */
+export const EnvTypes = Env;
