--- src/core/interfaces/EnvTypes.ts
@@ -1,13 +1,0 @@
-/**
- * @module common/core
- */ /** */
-
-import {Env} from "./Env";
-/**
- *
- */
-export class EnvTypes {
-    static PROD: Env = "production";
-    static DEV: Env = "development";
-    static TEST: Env = "test";
-}
