+++ src/core/interfaces/index.ts
@@ -4,4 +4,3 @@
 export * from "./Type";
 export * from "./DecoratorParameters";
 export * from "./Env";
-export * from "./EnvTypes";
