+++ src/index.ts
@@ -8,8 +8,10 @@ import "reflect-metadata";
 
 export * from "./core";
 export * from "./di";
+export * from "./config";
 export * from "./jsonschema";
-export * from "./mvc";
 export * from "./converters";
+
 export * from "./filters";
+export * from "./mvc";
 export * from "./server";
