+++ src/mvc/interfaces/index.ts
@@ -10,10 +10,10 @@ export * from "./IMiddlewareError";
 export * from "./IMiddlewareOptions";
 export * from "./IMiddlewareProvider";
 export * from "./IControllerMiddlewares";
-export * from "./IRouterOptions";
+export * from "../../config/interfaces/IRouterOptions";
 export * from "./MiddlewareType";
-export * from "./IRouterOptions";
-export * from "./IRouterOptions";
+export * from "../../config/interfaces/IRouterOptions";
+export * from "../../config/interfaces/IRouterOptions";
 export * from "./PathParamsType";
 export * from "./IChildrenController";
 export * from "./IResponseOptions";
