+++ src/server/index.ts
@@ -4,5 +4,4 @@
  */ /** */
 export * from "./interfaces";
 export * from "./components/ServerLoader";
-export * from "./services/ServerSettingsService";
 export * from "./decorators/serverSettings";
