+++ src/server/interfaces/index.ts
@@ -7,6 +7,5 @@ export * from "./BeforeRoutesInit";
 export * from "./IComponentScanned";
 export * from "./OnRoutesInit";
 export * from "./OnServerReady";
-export * from "./IServerSettings";
 export * from "./IServerLifeCycle";
 export * from "./IHTTPSServerOptions";
