+++ src/testing/loadInjector.ts
@@ -3,11 +3,11 @@
  */
 /** */
 
-import {EnvTypes} from "../core/interfaces";
+import {globalServerSettings} from "../config";
+import {Env} from "../core/interfaces";
 import {ExpressApplication} from "../core/services/ExpressApplication";
 import {InjectorService} from "../di/services/InjectorService";
-import {ServerSettingsProvider} from "../server/class/ServerSettingsProvider";
-import {ServerSettingsService} from "../server/services/ServerSettingsService";
+
 export function loadInjector() {
     if (!InjectorService.has(ExpressApplication)) {
         /* istanbul ignore next */
@@ -16,15 +16,7 @@ export function loadInjector() {
             get: () => (app)
         };
         InjectorService.set(ExpressApplication, app);
-
-        /* istanbul ignore else */
-        if (!InjectorService.has(ServerSettingsService)) {
-
-            const settingsProvider = new ServerSettingsProvider();
-            settingsProvider.env = EnvTypes.TEST;
-
-            InjectorService.set(ServerSettingsService, settingsProvider.$get());
-        }
+        globalServerSettings.env = Env.TEST;
     }
 
     InjectorService.load();
